storage move final metric structs into the metrics module
